requirements.txt
openai>=1.0.0
langgraph~=0.6.1
langchain~=0.3.27
langchain-core>=0.1.40
langchain-google-genai~=2.0.10
google-generativeai~=0.8.5
langchain-experimental~=0.3.4
yfinance>=0.2.65
requests>=2.31.0
beautifulsoup4>=4.13.0
python-dotenv>=1.0.0
tiktoken>=0.9.0
pygooglenews~=0.1.3
newspaper3k~=0.2.8
lxml[html_clean]>=5.2.0
finnhub-python~=2.4.24
tabulate~=0.9.0
praw==7.8.1
tifffile>=2023.8
shapely>=2.0

pandas~=2.3.1
pydantic~=2.11.7
numpy~=2.3.2
matplotlib~=3.10.6